decouple replacePhpDocTypeByAnother() to PhpDocModifier
